Update some test builders
